Create categories page - create categories component
