+++ src/components/Categories.tsx
@@ -1,0 +1,20 @@
+import React, { FC } from 'react';
+import CategoryCard, { CategoryProps } from './CategoryCard';
+
+const Categories: FC<{ categories: CategoryProps[] }> = ({ categories }) => {
+	return (
+		<div
+			style={{
+				display: 'grid',
+				gridTemplateColumns: 'repeat(auto-fit,minmax(260px,1fr))',
+				gap: '1rem',
+			}}
+		>
+			{categories.map((category, index) => (
+				<CategoryCard category={category} key={`category_${index}`} />
+			))}
+		</div>
+	);
+};
+
+export default Categories;
