--- src/pages/categories/composite.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const composite = () => {
-	return <div>composite</div>;
-};
-
-export default composite;
